Rename `StandardMapping` to `PredefinedMapping`
